Update to only show habit for dates after creation
Checklist now only displays checkboxes after a habit's creation date.
Grid now only counts habits in denominator if it were created before
the date of that grid square.
Habit methods to check and get completion are changed to use a
CompletionStatus Enum that adds the additional state of being inactive
in addition to the original complete/incomplete.